+++ coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/issued_coupon/IssuedCouponStoreImpl.kt
@@ -3,7 +3,7 @@ package coupon_platform.infrastructure.issued_coupon
 import coupon_platform.domain.issued_coupon.entitiy.IssuedCoupon
 import coupon_platform.domain.issued_coupon.repository.IssuedCouponStore
 import coupon_platform.infrastructure.cache.global_redis.handler.RedisHandlerOfIssuedCoupon
-import coupon_platform.infrastructure.cache.global_redis.util.RedisTTLConstants.ISSUED_COUPON_KEY_TTL
+import coupon_platform.infrastructure.cache.global_redis.util.CacheConstants.ISSUED_COUPON_KEY_TTL
 import org.springframework.stereotype.Repository
 
 @Repository
